First step towards rebranding the agreement
This work is based on the firefox distribution agreement taken from here:
https://extensionworkshop.com/documentation/publish/firefox-add-on-distribution-agreement/

This first step is updating links where possible and replaces Mozilla by MZLA which I picked for a placeholder. Ryan used "Thunderbird Team" in the other document, which might fit as well. But someone with more legal insight needs to decide what is correct.

I also added some todos were more work is needed.

Approving this PR will not make it visible on DTN, as it will be merged into the "add-add-on-policies" branch. Once we have everything up and running, we can merge that into master to add them to DTN.