Decouple the blocking policy from the other policies, so it applies to all add-ons
The blocking process applies to all add-ons and is not limited to listed add-ons. The blocking policy should therefore be formulated in a general way and should not depend on the distribution agreement or the review policies.

The following sentence has been removed:

The Thunderbird Team only blocks add-ons based on the [Thunderbird add-on distribution agreement](thunderbird-add-on-distribution-agreement.md) and [add-on policies](review-policy-for-thunderbird-add-ons.md). When requesting a block, please read these carefully as the policies do allow certain forms of monetization and data collection.

because THIS document describes the blocking characteristics.