Update using-a-background-page.md
The original text tends to be understand as if the messageTools.mjs module is "global" whilst this should have been created by the extension developer and included in the extension project directory. Originally, this was not clear. Indeed, I did not understand right away and get error message in the console when I tried to add the extension in Thunderbird.