Improve description of bootstrap.py in Linux build prerequisites
The description of the `bootstrap.py` script felt surprising to me because it starts by mentioning all the steps it does before mentioning the script itself.

It seems to be written from the perspective of the author of the script ("we have written…") rather than from the perspective of a new contributor, who doesn't know what those steps are in the first place. The wording I am proposing feels more natural in that respect.